Fix asmdef cycle: move SystemDrawer types; play mode test fixes
Retarget assemblies (BedogaGenerator/SystemDrawer/Narrative), add playback
interfaces in Locomotion.Runtime, relocate Spatial4DServiceWizard and animator
services. PhysicsCardSolver SolveForGoal prefers goal-matched cards without a
temporal graph; SectionLimits null-safe; Narrative JSON Vector3 converter;
ThoughtSimilarity camel-case tokens; tests updated.

Made-with: Cursor

+++ Assets/locomotion/Tests/NarrativeExecutorAndSchedulerTests.cs
@@ -311,7 +311,7 @@ public class NarrativeExecutorAndSchedulerTests
         };
 
         var result = action.Execute(ctx, state);
-        Assert.AreEqual(BehaviorTreeStatus.Success, result);
+        Assert.AreEqual((int)BehaviorTreeStatus.Success, (int)result);
 
         var azField = lightingType.GetField("sunAzimuthDeg");
         var sunVisibleField = lightingType.GetField("sunVisible");
